Testing OpenWeather API with lat and lon. It works. Now I need to make Retrofit work with multiple API's

+++ app/src/main/java/com/company/elverano/data/openWeather/OpenWeatherDaily.kt
@@ -19,7 +19,7 @@ data class OpenWeatherDaily(
     val wind_gust: Double = 0.0,
     val weather: ArrayList<OpenWeatherCurrent.OpenWeatherResult>,
     val clouds: Int = 0,
-    val pop: Int = 0,
+    val pop: Double = 0.0,
     val uvi: Double = 0.0
 ) : Parcelable {
